+++ app/src/main/java/sopt/co/kr/millielibraryandroid/api/ServiceCreator.kt
@@ -1,0 +1,15 @@
+package sopt.co.kr.millielibraryandroid.api
+
+import retrofit2.Retrofit
+import retrofit2.converter.gson.GsonConverterFactory
+
+object ServiceCreator {
+    private const val BASE_URL = "3.36.64.41:5000/"
+
+    private val retrofit: Retrofit = Retrofit.Builder()
+        .baseUrl(BASE_URL)
+        .addConverterFactory(GsonConverterFactory.create())
+        .build()
+
+    val milliService: MilliService = retrofit.create(MilliService::class.java)
+}
