--- app/src/main/java/sopt/co/kr/millielibraryandroid/api/data/BookInfo.kt
@@ -1,10 +1,0 @@
-package sopt.co.kr.millielibraryandroid.api.data
-
-data class BookInfo(
-    val image: Int,
-    val bookName: String,
-    val bookWriter: String,
-    val highLightNumber: String,
-    val bookContent: String,
-    val bookDate: String
-)
